test(#3928): generalise the corrupt-byte fixture harness to a Mutation enum and add a first-partition-HEADER mutation
Renames support/corrupt_clustering_fixture.rs -> corrupt_byte_fixture.rs (the
harness no longer only corrupts a clustering value) and replaces the
needles/flip-offset pair on FixtureSpec with a Mutation enum, so the #3782 row
arm and the #3928 header arm share one staging + CRC + measurement path.

The header site needs no index component: a Cassandra Data.db begins with its
first partition's header, so the byte is derived from the fixture's own
decompressed bytes and ACCEPTED only after re-decompressing and observing that
EXACTLY the target position changed to exactly the wanted value.
